requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
Werkzeug>=3.0.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.16.4
moviepy==1.0.3
Pillow==9.5.0
requests==2.31.0
python-dotenv==1.0.0
openai>=1.12.0
pydantic>=2.6.0
python-jose>=3.3.0
tenacity>=8.2.0
pytest>=8.0.0
pytest-cov>=4.1.0
redis==5.0.1
edge-tts>=6.1.9
nltk>=3.8.1
python-magic>=0.4.27
